fix(comments): conditional rendering of `CommentsProvider`
