+++ packages/sanity/src/structure/panes/document/comments/index.ts
@@ -1,0 +1,1 @@
+export * from './CommentsWrapper'
